+++ sdk/attestation/Azure.Security.Attestation/src/Models/TpmAttestationResponse.cs
@@ -12,7 +12,6 @@ namespace Azure.Security.Attestation.Models
     /// </summary>
     public partial class TpmAttestationResponse
     {
-
         /// <summary>
         /// Attestation Response data. See  https://docs.microsoft.com/en-us/azure/attestation/virtualization-based-security-protocol for more details
         /// </summary>