+++ sdk/core/Azure.Core.TestFramework/src/TestEnvironment.cs
@@ -11,6 +11,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Azure.Identity;
 using System.ComponentModel;
+using System.Linq;
 
 namespace Azure.Core.TestFramework
 {
@@ -30,21 +31,34 @@ namespace Azure.Core.TestFramework
 
         private readonly Dictionary<string, string> _environmentFile = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase);
 
-        protected TestEnvironment(string serviceName)
+        protected TestEnvironment()
         {
-            _prefix = serviceName.ToUpperInvariant() + "_";
             if (RepositoryRoot == null)
             {
                 throw new InvalidOperationException("Unexpected error, repository root not found");
             }
 
-            var sdkDirectory = Path.Combine(RepositoryRoot, "sdk", serviceName);
-            if (!Directory.Exists(sdkDirectory))
+            var testProject = GetSourcePath(GetType().Assembly);
+            var sdkDirectory = Path.GetFullPath(Path.Combine(RepositoryRoot, "sdk"));
+            var serviceName = Path.GetFullPath(testProject)
+                .Substring(sdkDirectory.Length)
+                .Trim(Path.DirectorySeparatorChar)
+                .Split(Path.DirectorySeparatorChar).FirstOrDefault();
+
+            if (string.IsNullOrWhiteSpace(serviceName))
             {
-                throw new InvalidOperationException($"SDK directory {sdkDirectory} not found");
+                throw new InvalidOperationException($"Unable to determine the service name from test project path {testProject}");
             }
 
-            var testEnvironmentFile = Path.Combine(RepositoryRoot, "sdk", serviceName, "test-resources.json.env");
+            var serviceSdkDirectory = Path.Combine(sdkDirectory, serviceName);
+            if (!Directory.Exists(sdkDirectory))
+            {
+                throw new InvalidOperationException($"SDK directory {serviceSdkDirectory} not found");
+            }
+
+            _prefix = serviceName.ToUpperInvariant() + "_";
+
+            var testEnvironmentFile = Path.Combine(serviceSdkDirectory, "test-resources.json.env");
             if (File.Exists(testEnvironmentFile))
             {
                 var json = JsonDocument.Parse(
@@ -289,6 +303,18 @@ namespace Azure.Core.TestFramework
             }
 
             return _recording.GetVariable(name, null);
+        }
+
+        internal static string GetSourcePath(Assembly assembly)
+        {
+            if (assembly == null) throw new ArgumentNullException(nameof(assembly));
+
+            var testProject = assembly.GetCustomAttributes<AssemblyMetadataAttribute>().Single(a => a.Key == "SourcePath").Value;
+            if (string.IsNullOrEmpty(testProject))
+            {
+                throw new InvalidOperationException($"Unable to determine the test directory for {assembly}");
+            }
+            return testProject;
         }
     }
 }