+++ sdk/eventhub/Azure.Messaging.EventHubs.Shared/src/BlobCheckpointStore/BlobsCheckpointStore.cs
@@ -5,6 +5,8 @@ using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
+using System.Linq;
+using System.Text.Json;
 using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
@@ -14,6 +16,7 @@ using Azure.Messaging.EventHubs.Core;
 using Azure.Messaging.EventHubs.Primitives;
 using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Models;
+using Azure.Storage.Blobs.Specialized;
 
 namespace Azure.Messaging.EventHubs.Processor
 {
@@ -42,6 +45,13 @@ namespace Azure.Messaging.EventHubs.Processor
         private const string CheckpointPrefix = "{0}/{1}/{2}/checkpoint/";
 
         /// <summary>
+        ///   Specifies a string that filters the results to return only legacy checkpoint blobs whose name begins
+        ///   with the specified prefix.
+        /// </summary>
+        ///
+        private const string LegacyCheckpointPrefix = "{0}/{1}/{2}/";
+
+        /// <summary>
         ///   Specifies a string that filters the results to return only ownership blobs whose name begins
         ///   with the specified prefix.
         /// </summary>
@@ -62,20 +72,29 @@ namespace Azure.Messaging.EventHubs.Processor
         private EventHubsRetryPolicy RetryPolicy { get; }
 
         /// <summary>
+        ///   Indicates whether to read legacy checkpoints when no current version checkpoints are available.
+        /// </summary>
+        ///
+        private bool InitializeWithLegacyCheckpoints { get; }
+
+        /// <summary>
         ///   Initializes a new instance of the <see cref="BlobsCheckpointStore" /> class.
         /// </summary>
         ///
         /// <param name="blobContainerClient">The client used to interact with the Azure Blob Storage service.</param>
         /// <param name="retryPolicy">The retry policy to use as the basis for interacting with the Storage Blobs service.</param>
+        /// <param name="initializeWithLegacyCheckpoints">Indicates whether to read legacy checkpoint when no current version checkpoint is available for a partition.</param>
         ///
         public BlobsCheckpointStore(BlobContainerClient blobContainerClient,
-                                    EventHubsRetryPolicy retryPolicy)
+                                    EventHubsRetryPolicy retryPolicy,
+                                    bool initializeWithLegacyCheckpoints = false)
         {
             Argument.AssertNotNull(blobContainerClient, nameof(blobContainerClient));
             Argument.AssertNotNull(retryPolicy, nameof(retryPolicy));
 
             ContainerClient = blobContainerClient;
             RetryPolicy = retryPolicy;
+            InitializeWithLegacyCheckpoints = initializeWithLegacyCheckpoints;
             BlobsCheckpointStoreCreated(nameof(BlobsCheckpointStore), blobContainerClient.AccountName, blobContainerClient.Name);
         }
 
@@ -275,11 +294,9 @@ namespace Azure.Messaging.EventHubs.Processor
             cancellationToken.ThrowIfCancellationRequested<TaskCanceledException>();
             ListCheckpointsStart(fullyQualifiedNamespace, eventHubName, consumerGroup);
 
-            var prefix = string.Format(CultureInfo.InvariantCulture, CheckpointPrefix, fullyQualifiedNamespace.ToLowerInvariant(), eventHubName.ToLowerInvariant(), consumerGroup.ToLowerInvariant());
-            var checkpointCount = 0;
-
-            async Task<IEnumerable<EventProcessorCheckpoint>> listCheckpointsAsync(CancellationToken listCheckpointsToken)
+            async Task<List<EventProcessorCheckpoint>> listCheckpointsAsync(CancellationToken listCheckpointsToken)
             {
+                var prefix = string.Format(CultureInfo.InvariantCulture, CheckpointPrefix, fullyQualifiedNamespace.ToLowerInvariant(), eventHubName.ToLowerInvariant(), consumerGroup.ToLowerInvariant());
                 var checkpoints = new List<EventProcessorCheckpoint>();
 
                 await foreach (BlobItem blob in ContainerClient.GetBlobsAsync(traits: BlobTraits.Metadata, prefix: prefix, cancellationToken: listCheckpointsToken).ConfigureAwait(false))
@@ -322,13 +339,82 @@ namespace Azure.Messaging.EventHubs.Processor
                     }
                 }
 
-                checkpointCount = checkpoints.Count;
                 return checkpoints;
             };
 
+            async Task<List<EventProcessorCheckpoint>> listLegacyCheckpointsAsync(List<EventProcessorCheckpoint> existingCheckpoints, CancellationToken listCheckpointsToken)
+            {
+                // Legacy checkpoints are not normalized to lowercase
+                var legacyPrefix = string.Format(CultureInfo.InvariantCulture, LegacyCheckpointPrefix, fullyQualifiedNamespace, eventHubName, consumerGroup);
+                var checkpoints = new List<EventProcessorCheckpoint>();
+
+                await foreach (BlobItem blob in ContainerClient.GetBlobsAsync(prefix: legacyPrefix, cancellationToken: listCheckpointsToken).ConfigureAwait(false))
+                {
+                    // Skip new checkpoints and empty blobs
+                    if (blob.Properties.ContentLength == 0)
+                    {
+                        continue;
+                    }
+
+                    var partitionId = blob.Name.Substring(legacyPrefix.Length);
+
+                    // Check whether there is already a checkpoint for this partition id
+                    if (existingCheckpoints.Any(existingCheckpoint => string.Equals(existingCheckpoint.PartitionId, partitionId, StringComparison.Ordinal)))
+                    {
+                        continue;
+                    }
+
+                    var startingPosition = default(EventPosition?);
+
+                    BlobBaseClient blobClient = ContainerClient.GetBlobClient(blob.Name);
+                    using var memoryStream = new MemoryStream();
+                    await blobClient.DownloadToAsync(memoryStream, listCheckpointsToken).ConfigureAwait(false);
+
+                    TryReadLegacyCheckpoint(
+                        memoryStream.GetBuffer().AsSpan(0, (int)memoryStream.Length),
+                        out long? offset,
+                        out long? sequenceNumber);
+
+                    if (offset.HasValue)
+                    {
+                        startingPosition = EventPosition.FromOffset(offset.Value, false);
+                    }
+                    else if (sequenceNumber.HasValue)
+                    {
+                        startingPosition = EventPosition.FromSequenceNumber(sequenceNumber.Value, false);
+                    }
+
+                    if (startingPosition.HasValue)
+                    {
+                        checkpoints.Add(new BlobStorageCheckpoint
+                        {
+                            FullyQualifiedNamespace = fullyQualifiedNamespace,
+                            EventHubName = eventHubName,
+                            ConsumerGroup = consumerGroup,
+                            PartitionId = partitionId,
+                            StartingPosition = startingPosition.Value,
+                            Offset = offset,
+                            SequenceNumber = sequenceNumber
+                        });
+                    }
+                    else
+                    {
+                        InvalidCheckpointFound(partitionId, fullyQualifiedNamespace, eventHubName, consumerGroup);
+                    }
+                }
+
+                return checkpoints;
+            };
+
+            List<EventProcessorCheckpoint> checkpoints = null;
             try
             {
-                return await ApplyRetryPolicy(listCheckpointsAsync, cancellationToken).ConfigureAwait(false);
+                checkpoints = await ApplyRetryPolicy(listCheckpointsAsync, cancellationToken).ConfigureAwait(false);
+                if (InitializeWithLegacyCheckpoints)
+                {
+                    checkpoints.AddRange(await ApplyRetryPolicy(ct => listLegacyCheckpointsAsync(checkpoints, ct), cancellationToken).ConfigureAwait(false));
+                }
+                return checkpoints;
             }
             catch (RequestFailedException ex) when (ex.ErrorCode == BlobErrorCode.ContainerNotFound)
             {
@@ -342,7 +428,7 @@ namespace Azure.Messaging.EventHubs.Processor
             }
             finally
             {
-                ListCheckpointsComplete(fullyQualifiedNamespace, eventHubName, consumerGroup, checkpointCount);
+                ListCheckpointsComplete(fullyQualifiedNamespace, eventHubName, consumerGroup, checkpoints?.Count ?? 0);
             }
         }
 
@@ -491,6 +577,71 @@ namespace Azure.Messaging.EventHubs.Processor
 
             await ApplyRetryPolicy(wrapper, cancellationToken).ConfigureAwait(false);
             return result;
+        }
+
+        /// <summary>
+        ///   Attempts to read a legacy checkpoint JSON format and extract an offset and a sequence number
+        /// </summary>
+        ///
+        /// <param name="data">The binary representation of the checkpoint JSON.</param>
+        /// <param name="offset">The parsed offset. null if not found.</param>
+        /// <param name="sequenceNumber">The parsed sequence number. null if not found.</param>
+        ///
+        /// <remarks>
+        ///   Sample checkpoint JSON:
+        ///   {
+        ///       "PartitionId":"0",
+        ///       "Owner":"681d365b-de1b-4288-9733-76294e17daf0",
+        ///       "Token":"2d0c4276-827d-4ca4-a345-729caeca3b82",
+        ///       "Epoch":386,
+        ///       "Offset":"8591964920",
+        ///       "SequenceNumber":960180
+        ///   }
+        /// /// </remarks>
+        private static void TryReadLegacyCheckpoint(Span<byte> data, out long? offset, out long? sequenceNumber)
+        {
+            offset = null;
+            sequenceNumber = null;
+
+            var jsonReader = new Utf8JsonReader(data);
+
+            try
+            {
+                if (!jsonReader.Read() || jsonReader.TokenType != JsonTokenType.StartObject) return;
+
+                while (jsonReader.Read() && jsonReader.TokenType == JsonTokenType.PropertyName)
+                {
+                    switch (jsonReader.GetString())
+                    {
+                        case "Offset":
+                            if (!jsonReader.Read() ||
+                                jsonReader.GetString() is not string offsetString ||
+                                !long.TryParse(offsetString, out long offsetValue))
+                            {
+                                return;
+                            }
+
+                            offset = offsetValue;
+                            break;
+                        case "SequenceNumber":
+                            if (!jsonReader.Read() ||
+                                !jsonReader.TryGetInt64(out long sequenceNumberValue))
+                            {
+                                return;
+                            }
+
+                            sequenceNumber = sequenceNumberValue;
+                            break;
+                        default:
+                            jsonReader.Skip();
+                            break;
+                    }
+                }
+            }
+            catch (JsonException)
+            {
+                // Ignore this because if the data is malformed, it will be treated as if the checkpoint didn't exist.
+            }
         }
 
         /// <summary>