+++ sdk/eventhub/Microsoft.Azure.WebJobs.Extensions.EventHubs/src/Listeners/EventHubListener.cs
@@ -23,7 +23,6 @@ namespace Microsoft.Azure.WebJobs.EventHubs
 {
     internal sealed class EventHubListener : IListener, IEventProcessorFactory, IScaleMonitorProvider
     {
-        private static readonly Dictionary<string, object> EmptyScope = new Dictionary<string, object>();
         private readonly ITriggeredFunctionExecutor _executor;
         private readonly EventProcessorHost _eventProcessorHost;
         private readonly bool _singleDispatch;
@@ -133,9 +132,11 @@ namespace Microsoft.Azure.WebJobs.EventHubs
 
             public async Task ProcessEventsAsync(EventProcessorHostPartition context, IEnumerable<EventData> messages)
             {
+                var events = messages.ToArray();
+
                 var triggerInput = new EventHubTriggerInput
                 {
-                    Events = messages.ToArray(),
+                    Events = events,
                     PartitionContext = context
                 };
 
@@ -159,7 +160,7 @@ namespace Microsoft.Azure.WebJobs.EventHubs
                             TriggerDetails = eventHubTriggerInput.GetTriggerDetails(context)
                         };
 
-                        Task task = TryExecuteWithLoggingAsync(input, triggerInput.Events[i]);
+                        Task task = _executor.TryExecuteAsync(input, _cts.Token);
                         invocationTasks.Add(task);
                     }
 
@@ -178,10 +179,7 @@ namespace Microsoft.Azure.WebJobs.EventHubs
                         TriggerDetails = triggerInput.GetTriggerDetails(context)
                     };
 
-                    using (_logger.BeginScope(GetLinksScope(triggerInput.Events)))
-                    {
-                        await _executor.TryExecuteAsync(input, _cts.Token).ConfigureAwait(false);
-                    }
+                    await _executor.TryExecuteAsync(input, _cts.Token).ConfigureAwait(false);
                 }
 
                 // Checkpoint if we processed any events.
@@ -191,17 +189,9 @@ namespace Microsoft.Azure.WebJobs.EventHubs
                 // so that is the responsibility of the user's function currently. E.g.
                 // the function should have try/catch handling around all event processing
                 // code, and capture/log/persist failed events, since they won't be retried.
-                if (messages.Any())
+                if (events.Any())
                 {
-                    await CheckpointAsync(messages.Last(), context).ConfigureAwait(false);
-                }
-            }
-
-            private async Task TryExecuteWithLoggingAsync(TriggeredFunctionData input, EventData message)
-            {
-                using (_logger.BeginScope(GetLinksScope(message)))
-                {
-                    await _executor.TryExecuteAsync(input, _cts.Token).ConfigureAwait(false);
+                    await CheckpointAsync(events.Last(), context).ConfigureAwait(false);
                 }
             }
 
@@ -245,56 +235,6 @@ namespace Microsoft.Azure.WebJobs.EventHubs
             public void Dispose()
             {
                 Dispose(true);
-            }
-
-            private static Dictionary<string, object> GetLinksScope(EventData message)
-            {
-                if (TryGetLinkedActivity(message, out var link))
-                {
-                    return new Dictionary<string, object> {["Links"] = new[] {link}};
-                }
-
-                return EmptyScope;
-            }
-
-            private static Dictionary<string, object> GetLinksScope(EventData[] messages)
-            {
-                List<Activity> links = null;
-
-                foreach (var message in messages)
-                {
-                    if (TryGetLinkedActivity(message, out var link))
-                    {
-                        if (links == null)
-                        {
-                            links = new List<Activity>(messages.Length);
-                        }
-
-                        links.Add(link);
-                    }
-                }
-
-                if (links != null)
-                {
-                    return new Dictionary<string, object> {["Links"] = links};
-                }
-
-                return EmptyScope;
-            }
-
-            private static bool TryGetLinkedActivity(EventData message, out Activity link)
-            {
-                link = null;
-
-                if (((message.SystemProperties != null && message.SystemProperties.TryGetValue("Diagnostic-Id", out var diagnosticIdObj)) || message.Properties.TryGetValue("Diagnostic-Id", out diagnosticIdObj))
-                    && diagnosticIdObj is string diagnosticIdString)
-                {
-                    link = new Activity("Microsoft.Azure.EventHubs.Process");
-                    link.SetParentId(diagnosticIdString);
-                    return true;
-                }
-
-                return false;
             }
 
             private static string GetOperationDetails(EventProcessorHostPartition context, string operation)