+++ sdk/synapse/Azure.Analytics.Synapse.Spark/tests/samples/Sample2_ExecuteSparkStatement.cs
@@ -16,19 +16,17 @@ namespace Azure.Analytics.Synapse.Samples
     public partial class ExecuteSparkStatement
     {
         [Test]
+        [Ignore("https://github.com/Azure/azure-sdk-for-net/issues/17455")]
         public void ExecuteSparkStatementSync()
         {
             // Environment variable with the Synapse workspace endpoint.
-            string workspaceUrl = TestEnvironment.WorkspaceUrl;
+            string endpoint = TestEnvironment.EndpointUrl;
 
             // Environment variable with the Synapse Spark pool name.
             string sparkPoolName = TestEnvironment.SparkPoolName;
 
-            #region Snippet:SparkSessionSample1SparkSessionClient
-            SparkSessionClient client = new SparkSessionClient(new Uri(workspaceUrl), sparkPoolName, new DefaultAzureCredential());
-            #endregion
+            SparkSessionClient client = new SparkSessionClient(new Uri(endpoint), sparkPoolName, new DefaultAzureCredential());
 
-            #region Snippet:SparkSessionSample1StartSparkSession
             SparkSessionOptions request = new SparkSessionOptions(name: $"session-{Guid.NewGuid()}")
             {
                 DriverMemory = "28g",
@@ -39,35 +37,24 @@ namespace Azure.Analytics.Synapse.Samples
             };
 
             SparkSession sessionCreated = client.CreateSparkSession(request);
-            #endregion
 
-            #region Snippet:SparkSessionSample1GetSparkSession
             SparkSession session = client.GetSparkSession(sessionCreated.Id);
             Debug.WriteLine($"Session is returned with name {session.Name} and state {session.State}");
-            #endregion
 
-            #region Snippet:SparkSessionSample1ExecuteSparkStatement
             SparkStatementOptions sparkStatementRequest = new SparkStatementOptions
             {
                 Kind = SparkStatementLanguageType.Spark,
                 Code = @"print(""Hello world\n"")"
             };
             SparkStatement statementCreated = client.CreateSparkStatement(sessionCreated.Id, sparkStatementRequest);
-            #endregion
 
-            #region Snippet:SparkSessionSample1GetSparkStatement
             SparkStatement statement = client.GetSparkStatement(sessionCreated.Id, statementCreated.Id);
             Debug.WriteLine($"Statement is returned with id {statement.Id} and state {statement.State}");
-            #endregion
 
-            #region Snippet:SparkSessionSample1CancelSparkStatement
             SparkStatementCancellationResult cancellationResult = client.CancelSparkStatement(sessionCreated.Id, statementCreated.Id);
             Debug.WriteLine($"Statement is cancelled with message {cancellationResult.Msg}");
-            #endregion
 
-            #region Snippet:SparkSessionSample1StopSparkSession
             Response operation = client.CancelSparkSession(sessionCreated.Id);
-            #endregion
         }
     }
 }