+++ sdk/core/Azure.Core/CHANGELOG.md
@@ -1,11 +1,12 @@
 # Release History
 
-## 1.7.0-beta.1 (Unreleased)
+## 1.7.0 (2020-12-14)
 
-### Added
+### New Features
 - `System.Text.Json.JsonConverter` implementation for the `ETag`
+- Synchronous `HttpClient` support on .NET 5.
 
-### Fixes
+### Key Bug Fixes
 - System proxy settings are correctly applied on .NET Framework
 
 ## 1.6.0 (2020-10-28)